requirements.txt
python-telegram-bot==12.7
google==2.0.3
requests==2.22.0
pyowm==2.10.0
wikipedia==1.4.0